requirements.txt
absl-py==2.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.21.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.0.1
fastapi==0.115.14
fastjsonschema==2.20.0
filelock==3.18.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.58.5
fqdn==1.5.1
fsspec==2025.5.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.174.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.14.0
htmldocx==0.0.6
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.33.0
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.1
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.10.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
libclang==18.1.1
lxml==5.4.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml_dtypes==0.5.1
MouseInfo==0.1.3
mpmath==1.3.0
murmurhash==1.0.13
namex==0.1.0
narwhals==1.44.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.1.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
overrides==7.7.0
packaging==24.1
pandas==2.3.0
pandocfilters==1.5.1
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.0.0
platformdirs==4.2.2
plotly==6.2.0
preshed==3.0.10
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyMsgBox==1.0.9
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==2.0.7
python-multipart==0.0.20
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.19.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==71.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.0
streamlit-option-menu==0.4.0
sympy==1.14.0
tabula-py==2.10.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminado==0.18.1
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.3.0
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.16.0
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.35.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wheel==0.45.1
widgetsnbextension==4.0.11
wikipedia==1.4.0
wrapt==1.17.2
xlrd==2.0.2
